events/requirements.txt
flask==3.0.0

python-dotenv==1.0.0
sqlalchemy==2.0.36
pymysql==1.1.0
cryptography==41.0.7
pyjwt==2.8.0
requests==2.31.0
